stream: Divide write timeout by 2
When setting the socket's write timeout, it will instead timeout after
double the time set. In order to get the correct behavior, the timeout
is divided by 2.
We have no idea why it behaves this way, but there is a discussion on
stack overflow mentioning it:
https://stackoverflow.com/questions/26298040/socket-so-rcvtimeo-timeout-is-double-the-set-value-in-c-vc?noredirect=1&lq=1

Note that this is a local patch and should not be merged upstream.